pyproject.toml
[project]
name = "backend-randomapi"
version = "0.1.0"
description = "\"бекенд мини приложения в ТГ РАНДОМАЙЗЕР АПИ\""
authors = [
    {name = "Ury Dol3wsky",email = "dol3wsky@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "load-dotenv (>=0.1.0,<0.2.0)",
    "aiohttp (>=3.11.14,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"

